LPD-57959 Inline PhraseSuggesterTranslator/PhraseSuggesterTranslatorImpl and TermSuggesterTranslator/TermSuggesterTranslatorImpl
